Decouple queries & mutations from authentication
