--- backend/app/adapter/graphql/resolver/User.go
@@ -1,15 +1,0 @@
-package resolver
-
-import (
-	"short/app/entity"
-)
-
-// User stores user information used by GraphQL API
-type User struct {
-	user entity.User
-}
-
-// Email retrieves user's email address
-func (u User) Email() *string {
-	return &u.user.Email
-}
