+++ backend/app/adapter/graphql/resolver/authmutation_test.go
@@ -1,0 +1,7 @@
+package resolver
+
+import "testing"
+
+func TestAuthMutation_CreateURL(t *testing.T) {
+
+}
