+++ backend/app/adapter/graphql/resolver/query.go
@@ -1,10 +1,8 @@
 package resolver
 
 import (
-	"short/app/adapter/graphql/scalar"
 	"short/app/usecase/auth"
 	"short/app/usecase/url"
-	"time"
 
 	"github.com/byliuyang/app/fw"
 )
@@ -32,47 +30,18 @@ func NewQuery(
 	}
 }
 
-// ViewerArgs represents possible arguments for viewer endpoint
-type ViewerArgs struct {
+// ViewerQueryArgs represents possible arguments for viewer endpoint
+type ViewerQueryArgs struct {
 	AuthToken *string
 }
 
-// Viewer extracts user information from authentication token
-func (q Query) Viewer(args *ViewerArgs) (*User, error) {
-	if args.AuthToken == nil {
-		return nil, nil
-	}
-
-	authToken := *args.AuthToken
-	user, err := q.authenticator.GetUser(authToken)
+// AuthQuery extracts user information from authentication token
+func (q Query) AuthQuery(args *ViewerQueryArgs) (*AuthQuery, error) {
+	user, err := viewer(args.AuthToken, q.authenticator)
 	if err != nil {
 		return nil, err
 	}
-	return &User{
-		user: user,
-	}, nil
-}
 
-// URLArgs represents possible argument for URL endpoint
-type URLArgs struct {
-	Alias       string
-	ExpireAfter *scalar.Time
-}
-
-// URL retrieves an URL persistent storage given alias and expiration time.
-func (q Query) URL(args *URLArgs) (*URL, error) {
-	trace := q.tracer.BeginTrace("Query.URL")
-	defer trace.End()
-
-	var expireAt *time.Time
-	if args.ExpireAfter != nil {
-		expireAt = &args.ExpireAfter.Time
-	}
-
-	u, err := q.urlRetriever.GetURL(args.Alias, expireAt)
-	if err != nil {
-		q.logger.Error(err)
-		return nil, err
-	}
-	return &URL{url: u}, nil
+	authQuery := NewAuthQuery(user, q.urlRetriever)
+	return &authQuery, nil
 }
