+++ backend/app/adapter/graphql/resolver/url.go
@@ -24,3 +24,7 @@ func (u URL) ExpireAt() *scalar.Time {
 
 	return &scalar.Time{Time: *u.url.ExpireAt}
 }
+
+func NewURL(url entity.URL) URL {
+	return URL{url: url}
+}
